mc1n2: Allow swapping of microphone ADC channels
The SGH-I777 libaudio has the configuration for
ADC0[0] and ADC0[1] swapped.

When using GT-I9100 libaudio, this results in
the microphones being improperly swapped between
the main mic and the NR mic near the earpiece.

This patch allows the mic channels to be swapped
within the driver to deal with incompatible libaudio.

It is configurable via
CONFIG_SND_SOC_MC1N2_MIC_ADC_SWAP in the defconfig
in case the SGH-I777 libaudio becomes available.

Change-Id: Id5738369f3614c6ccf710ac02ae96ee65eb4398b